+++ src/app/mod.rs
@@ -10,3 +10,4 @@ pub(crate) mod remove;
 pub(crate) mod rename;
 pub(crate) mod shell;
 pub(crate) mod sort;
+pub(crate) mod sync;